# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VPS Billing Platform API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Create a user account
      description: |
        Does not sign the account in. The session is issued by /auth/login, so there is
        one place that issues credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RegisterRequest"}
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [user]
                        properties:
                          user: {$ref: "#/components/schemas/User"}
        "409":
          description: The address is already registered
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "422":
          description: A field is invalid
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "429":
          description: Too many attempts from this address
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /auth/login:
    post:
      operationId: login
      summary: Sign in as a user
      description: |
        On success the session cookie is set and the CSRF token is returned in the body.
        The token cannot be read from the cookie, which is HttpOnly, so the body is the
        only way a client can obtain it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: Signed in
          headers:
            Set-Cookie:
              description: The session cookie. HttpOnly, SameSite=Lax, Secure outside development.
              schema: {type: string}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data: {$ref: "#/components/schemas/UserSession"}
        "401":
          description: |
            Invalid credentials. Returned identically for a wrong password and an unknown
            address, so the endpoint does not disclose whether an account exists.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: The account is suspended
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "429":
          description: Too many attempts from this address or against this account
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /auth/logout:
    post:
      operationId: logout
      summary: End the current user session
      description: Requires the X-CSRF-Token header of the session being ended.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204": {description: Signed out}
        "401":
          description: No session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: The CSRF token is missing or wrong
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: The authenticated user
      description: |
        Returns the CSRF token as well, which is how a reloaded page recovers it without a
        second endpoint.
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data: {$ref: "#/components/schemas/UserSession"}
        "401":
          description: No session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /admin/auth/login:
    post:
      operationId: adminLogin
      summary: Sign in as an administrator
      description: |
        Issues an administrator session in its own cookie. The customer and administrator
        credential spaces are separate in cookie name, storage namespace and subject type,
        so a customer session cannot authenticate an administrative endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: Signed in
          headers:
            Set-Cookie:
              description: The administrator session cookie.
              schema: {type: string}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data: {$ref: "#/components/schemas/AdminSession"}
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: The account is suspended
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "429":
          description: Too many attempts
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /admin/auth/logout:
    post:
      operationId: adminLogout
      summary: End the current administrator session
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204": {description: Signed out}
        "401":
          description: No session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: The CSRF token is missing or wrong
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /admin/auth/me:
    get:
      operationId: adminGetCurrent
      summary: The authenticated administrator, with effective permissions
      responses:
        "200":
          description: The current administrator
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data: {$ref: "#/components/schemas/AdminSession"}
        "401":
          description: No administrator session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: The account is suspended
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
  /orders:
    get:
      operationId: listOrders
      summary: The signed-in customer's orders
      responses:
        "200":
          description: Orders, most recent first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [orders]
                        properties:
                          orders:
                            type: array
                            items: {$ref: "#/components/schemas/Order"}
        "401":
          description: No session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    post:
      operationId: createOrder
      summary: Place an order
      description: |
        Prices the order from the catalogue. The caller states which plan and how many;
        it cannot state a price, because the unit price is read from the plan here.
        An invoice is opened with the order, before any payment exists.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [plan_id, quantity]
                    properties:
                      plan_id: {type: string, format: uuid}
                      quantity: {type: integer, minimum: 1, maximum: 1000}
                discount_minor: {type: integer, minimum: 0, default: 0}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [order]
                        properties:
                          order: {$ref: "#/components/schemas/Order"}
        "401":
          description: No session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: The CSRF token is missing or wrong
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "422":
          description: A field is invalid, or a plan is not on sale
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /orders/{id}:
    get:
      operationId: getOrder
      summary: One of the signed-in customer's orders
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [order]
                        properties:
                          order: {$ref: "#/components/schemas/Order"}
        "404":
          description: |
            No such order. The same answer whether it does not exist or belongs to
            somebody else, so one customer's identifiers cannot enumerate another's.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /orders/{id}/payments:
    post:
      operationId: startPayment
      summary: Ask a gateway to take the order's money
      description: |
        Returns the gateway's own payment identifier and the URL the customer completes
        the payment at. The idempotency key is derived from the order and the gateway, so
        a retry cannot create a second payment for the same order.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gateway]
              properties:
                gateway: {type: string}
      responses:
        "201":
          description: The payment was started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [payment]
                        properties:
                          payment: {$ref: "#/components/schemas/Payment"}
        "409":
          description: |
            The order is not payable, or a payment has already been started for it
            through this gateway.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /webhooks/payments/{gateway}:
    post:
      operationId: paymentWebhook
      summary: A gateway's callback
      description: |
        Public, because a gateway cannot hold a session. What authenticates it is the
        signature over the raw body, carried in the header the named gateway uses.
        The signature is verified before anything is looked up.
      parameters:
        - name: gateway
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
      responses:
        "200":
          description: |
            The notification was met. `data.settled` says whether this delivery was the
            one that performed the settlement, and `data.already_settled` says it had
            been performed earlier — both are success, because the caller's obligation
            was to deliver.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [settled, already_settled]
                        properties:
                          settled: {type: boolean}
                          already_settled: {type: boolean}
        "400":
          description: The body could not be parsed
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: The signature does not match
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "409":
          description: |
            The callback disagrees with the platform's record — a different amount or
            currency. Reported rather than absorbed, because the two sources believing
            they agree is the failure this endpoint exists to prevent.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /subscriptions:
    get:
      operationId: listSubscriptions
      summary: The signed-in customer's subscriptions
      responses:
        "200":
          description: Subscriptions, most recent first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [subscriptions]
                        properties:
                          subscriptions:
                            type: array
                            items: {$ref: "#/components/schemas/Subscription"}
        "401":
          description: No session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      summary: One of the customer's subscriptions
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: The subscription
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [subscription]
                        properties:
                          subscription: {$ref: "#/components/schemas/Subscription"}
        "404": {description: No such subscription}
  /subscriptions/{id}/cancel:
    post:
      operationId: cancelSubscription
      summary: Ask for the subscription to end when the period does
      description: |
        The current period has been paid for and is not cut short. The machine
        honours the request when it would next bill, instead of billing again.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: The request is recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [cancel_at_period_end]
                        properties:
                          cancel_at_period_end: {type: boolean, enum: [true]}
        "404": {description: No such subscription}
        "409": {description: The subscription has already ended}
  /subscriptions/{id}/renew:
    post:
      operationId: renewSubscription
      summary: Pay the due renewal now, from the wallet balance
      description: |
        The sweep would reach it on its own pass. The renewal is paid from the
        wallet, the period is extended in the same transaction, and the response
        carries the subscription in its new period.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: The subscription in its new period
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [subscription]
                        properties:
                          subscription: {$ref: "#/components/schemas/Subscription"}
        "404": {description: No such subscription}
        "409":
          description: The period has not ended, the subscription has ended, or the
            balance does not cover the renewal
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Order with items and payments}
        "404": {description: Not found}
...
  /admin/instances/{id}:
    get:
      operationId: adminGetInstance
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Instance detail with joins}
        "404": {description: Not found}
...
  /admin/tickets/{id}:
    get:
      operationId: adminGetTicket
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Ticket with messages}
        "404": {description: Not found}
  /admin/tickets/{id}/messages:
    post:
      operationId: adminReplyTicket
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: {type: string}
      responses:
        "201": {description: Message appended}
        "404": {description: Not found}
  /admin/tickets/{id}/close:
    post:
      operationId: adminCloseTicket
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Ticket closed}
        "404": {description: Not found}
...
  /admin/operations/{id}:
    get:
      operationId: getOperation
      summary: One operation and its steps
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: The operation's record
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OperationDetailResponse"}
        "403": {description: The administrator does not hold operations.read}
...
  /instances/{id}:
    get:
      operationId: getInstance
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": {description: Instance}
        "404": {description: Not found}
...
  /operations/{id}:
    get:
      operationId: getOperation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Operation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OperationResponse"}
...
components:
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, request_id]
      properties:
        success: {type: boolean, const: true}
        data: {type: object}
        request_id: {type: string}
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 12}
        locale:
          type: string
          enum: [zh-CN, en-US]
          description: Defaults to zh-CN when omitted.
        timezone: {type: string}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string}
        password: {type: string}
    User:
      type: object
      required: [id, email, status, locale, timezone]
      properties:
        id: {type: string, format: uuid}
        email: {type: string}
        status: {type: string, enum: [active, suspended]}
        locale: {type: string, enum: [zh-CN, en-US]}
        timezone: {type: string}
...
    UserSession:
      type: object
      required: [user, csrf_token, expires_at]
      properties:
        user: {$ref: "#/components/schemas/User"}
        csrf_token: {type: string}
        expires_at: {type: string, format: date-time}
    AdminSession:
      type: object
      required: [admin, csrf_token, expires_at]
      properties:
        admin: {$ref: "#/components/schemas/Admin"}
        csrf_token: {type: string}
        expires_at: {type: string, format: date-time}
...